move DrtRequest from [org.matsim.contrib.drt].data to .passenger
