+++ contribs/drt/src/main/java/org/matsim/contrib/drt/optimizer/insertion/DetourLinksProvider.java
@@ -27,7 +27,7 @@ import java.util.concurrent.ForkJoinPool;
 
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
-import org.matsim.contrib.drt.data.DrtRequest;
+import org.matsim.contrib.drt.passenger.DrtRequest;
 import org.matsim.contrib.drt.optimizer.VehicleData.Entry;
 import org.matsim.contrib.drt.optimizer.insertion.InsertionGenerator.Insertion;
 import org.matsim.contrib.drt.run.DrtConfigGroup;
