+++ contribs/drt/src/main/java/org/matsim/contrib/drt/optimizer/insertion/ParallelMultiVehicleInsertionProblem.java
@@ -28,7 +28,7 @@ import java.util.concurrent.ForkJoinPool;
 
 import org.apache.commons.math3.stat.descriptive.SummaryStatistics;
 import org.apache.log4j.Logger;
-import org.matsim.contrib.drt.data.DrtRequest;
+import org.matsim.contrib.drt.passenger.DrtRequest;
 import org.matsim.contrib.drt.optimizer.VehicleData.Entry;
 import org.matsim.contrib.drt.optimizer.insertion.DetourLinksProvider.DetourLinksSet;
 import org.matsim.contrib.drt.optimizer.insertion.InsertionGenerator.Insertion;
