+++ contribs/drt/src/main/java/org/matsim/contrib/drt/optimizer/insertion/PathDataProvider.java
@@ -18,7 +18,7 @@
 
 package org.matsim.contrib.drt.optimizer.insertion;
 
-import org.matsim.contrib.drt.data.DrtRequest;
+import org.matsim.contrib.drt.passenger.DrtRequest;
 import org.matsim.contrib.drt.optimizer.VehicleData;
 import org.matsim.contrib.dvrp.path.OneToManyPathSearch.PathData;
