+++ contribs/drt/src/main/java/org/matsim/contrib/drt/optimizer/insertion/SingleVehicleInsertionFilter.java
@@ -22,7 +22,7 @@ package org.matsim.contrib.drt.optimizer.insertion;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import org.matsim.contrib.drt.data.DrtRequest;
+import org.matsim.contrib.drt.passenger.DrtRequest;
 import org.matsim.contrib.drt.optimizer.VehicleData;
 import org.matsim.contrib.drt.optimizer.insertion.DetourTimesProvider.DetourTimesSet;
 import org.matsim.contrib.drt.optimizer.insertion.InsertionGenerator.Insertion;
