+++ contribs/drt/src/main/java/org/matsim/contrib/drt/optimizer/insertion/SingleVehicleInsertionProblem.java
@@ -22,7 +22,7 @@ package org.matsim.contrib.drt.optimizer.insertion;
 import java.util.List;
 import java.util.Optional;
 
-import org.matsim.contrib.drt.data.DrtRequest;
+import org.matsim.contrib.drt.passenger.DrtRequest;
 import org.matsim.contrib.drt.optimizer.VehicleData;
 import org.matsim.contrib.drt.optimizer.insertion.InsertionGenerator.Insertion;
 import org.matsim.contrib.drt.optimizer.insertion.PathDataProvider.PathDataSet;
