+++ contribs/drt/src/main/java/org/matsim/contrib/drt/optimizer/insertion/UnplannedRequestInserter.java
@@ -21,7 +21,7 @@ package org.matsim.contrib.drt.optimizer.insertion;
 
 import java.util.Collection;
 
-import org.matsim.contrib.drt.data.DrtRequest;
+import org.matsim.contrib.drt.passenger.DrtRequest;
 
 /**
  * @author michalm
