+++ contribs/drt/src/main/java/org/matsim/contrib/drt/schedule/DrtStopTask.java
@@ -24,7 +24,7 @@ import java.util.HashSet;
 import java.util.Set;
 
 import org.matsim.api.core.v01.network.Link;
-import org.matsim.contrib.drt.data.DrtRequest;
+import org.matsim.contrib.drt.passenger.DrtRequest;
 import org.matsim.contrib.dvrp.schedule.StayTaskImpl;
 
 /**
