+++ contribs/drt/src/main/java/org/matsim/contrib/drt/scheduler/RequestInsertionScheduler.java
@@ -22,7 +22,7 @@ package org.matsim.contrib.drt.scheduler;
 import java.util.List;
 
 import org.matsim.api.core.v01.network.Link;
-import org.matsim.contrib.drt.data.DrtRequest;
+import org.matsim.contrib.drt.passenger.DrtRequest;
 import org.matsim.contrib.drt.optimizer.VehicleData;
 import org.matsim.contrib.drt.optimizer.VehicleData.Stop;
 import org.matsim.contrib.drt.optimizer.insertion.InsertionWithPathData;
